ContinuedWorkingWithDatePicker
Changed to the onSelect function because it lets us work with the new value as opposed to the original date in the datePicker, the function currently takes the date as an input formats it and uses window.alert to display the data and clearly is the desired date range. The probelm still exists in getting data back from Datapicker to Dashboard.js, according to online documentation the Value prop is used for that but I cannot figure it out. Code still works as it did before, so nothing should be broken